Added project for xcode. Moved Visual Studio solution to projects folder.

+++ src/hash-functions.c
@@ -28,7 +28,7 @@
 /*
  *   Hash Functions
  */
-static __inline size_t int64_hash( uint64_t key )
+static __inline uint64_t int64_hash( uint64_t key )
 {
 	key = (~key) + (key << 21); /* key = (key << 21) - key - 1;*/
 	key = key ^ (key >> 24);
@@ -40,7 +40,7 @@ static __inline size_t int64_hash( uint64_t key )
 	return key;
 }
 
-static __inline size_t int32_hash( uint32_t a )
+static __inline uint32_t int32_hash( uint32_t a )
 {
 	a = (a+0x7ed55d16) + (a<<12);
 	a = (a^0xc761c23c) ^ (a>>19);
